Remove unused string escape from queryString. Fixes #55
If user is using SOQL, we assume they know what they are doing and will be careful with their string.

Sharing is on by default (via inherited) and there are two additional checks in place - Security.stripInaccessible and field.isAccessible.

Given that, I feel there are enough safety precautions to remove this artifact.